Fix EntityDeleteConfirmModalDisplay tests
