--- packages/components/src/internal/components/assay/constants.ts
@@ -1,13 +1,0 @@
-import { SCHEMAS } from '../../../internal/components/base/models/schemas';
-
-export const RUN_PROPERTIES_GRID_ID = 'assay-run-details';
-
-export const RUN_PROPERTIES_REQUIRED_COLUMNS = SCHEMAS.CBMB.concat(
-    'Name',
-    'RowId',
-    'ReplacesRun',
-    'ReplacedByRun',
-    'DataOutputs',
-    'DataOutputs/DataFileUrl',
-    'Batch'
-).toList();
