+++ packages/components/src/internal/components/entities/EntityDeleteConfirmModalDisplay.tsx
@@ -15,8 +15,8 @@
  */
 import React from 'react';
 
-import { ConfirmModal } from '../base/ConfirmModal';
-import { helpLinkNode } from '../../util/helpLinks';
+import { ConfirmModal } from '../../..';
+import { helpLinkNode } from '../../..';
 
 import { DeleteConfirmationData } from './actions';
 import { EntityDataType } from './models';
